feat: move formats/custom game to their own routes (#104)
* feat: move formats to standalone pages

* feat: move custom game to route

* chore: clean up game setup

* chore: clean up linting errors

* refactor: rename format component

* feat: add query to url for custom games
